Fix bug where we reported possible double initialization
We previously reported a possible double initialization because during
dataflow analysis, both TerminatorKind::Reattach and a generated statement
which assigned the spawn result to the destination would state that they'd
initialized the destination. This is illegal for an immutable variable, so
we saw two errors rather than the one error we expected.

The fix was to allow the assignment to handle all the work of propagating
dataflow information about the destination being assigned and allowing
Reattach to only perform control flow. Since Reattach itself shouldn't
be able to unwind on its own, this could simplify adding unwind / abort
support. Reattach now has no field for destination and is only concerned with
the continuation.