Add a module for marking Cilk tasks separately
mark_cilk_tasks builds a task tree and determines all reattach points for a given
task. This is useful when computing various dataflow analyses since sync changes
what variables are initialized (and other state, but that's harder to integrate
and not required to get code to compile).

We still have to use this module to correctly handle sync terminators of basic blocks
when finding what variables will be initialized, it would be nice if we had separate
notions of "will-be-synced" and "may-be-synced" for each corresponding kind of dataflow,
and we still need to integrate this with borrow-checking so that we don't kill loans
too early.