Update initialized-variable analysis to consider syncs
This commit extends mark_cilk_tasks::TaskTree with useful methods for
initialized variable analysis (mostly ways to observe the state of a
TaskTree). It primarily extends the analyses of initialized variables
to consider syncs as initializing all variables that are initialized
at reattachment points.

One part I'm not sure of: we want the number of initialized variables to
decrease when we merge in DefinitelyInitializedVariables if the reattaches
are from the same task, and increase if they're not from the same task.