Fix DefinitelyInitializedVariables to merge state correctly
This commit changes DefinitelyInitializedVariables to merge the
dataflow state at the places a task exits via join (intersection),
which makes sense because it reduces the number of initialized
places. We then use meet (union) to merge the initialized variable
state after all of the tasks are done at the sync. The bug with
syncing conditional spawns is still possible.