Use a preorder traversal rather than visitor to build TaskTree
When we used a visitor, we saw ICEs when building rustc. I think this is
because of the particular order of the visitor, but here we care about the
traversal order and only need to worry about basic block terminators.
A preorder traversal makes sense because we want to ensure that the only
block which has a new task constructed for it is the root of the Body,
and this is true in a preorder traversal as long as the Body is connected
(which seems to be a reasonable assumption).

If at some point the body is disconnected, we can allow reusing the default
task since the task we assign doesn't actually matter: the disconnected
portion of the graph (whichever one isn't reachable from the root) will
never be executed and should be removed by dead code elimination.