Avoid labeling basic blocks in unwind subgraph with a task
We now do not label basic blocks in the unwind subgraph with a task.
This is because the cleanup blocks in the unwind subgraph are reachable from
non-cleanup blocks when they unwind. This would lead to labeling unwind blocks
with many possible tasks, which breaks our invariant that blocks have exactly
one task.