Only inject FakeRead for when the RHS is not a CilkSpawn
Previously, the LHS of an assignment is always used by a FakeRead
for better diagnostics, since it's otherwise possible to create
variable that can't actually be used. The read makes those
initializations an error. However, the value is not available in
the case of a spawn until a sync, so we get this benefit anyways.