Add CilkScope to HIR
This commit adds CilkScope to hir::ExprKind. A future commit will
add CilkScope to thir::ExprKind.

This commit makes a few assumptions about the type of CilkScope.
It assumes that CilkScope evaluates to the type of the contained
block, which invites usages where a user uses the final value of a
CilkScope block. Due to existing rules around not using spawned values
without a sync (semantic restrictions around unused variables),
programs which contain scopes that evaluate to values composed of possibly-unsynced
values will see errors. This is because cilk_scope inserts its sync *after*
the last expression in the block rather than before it (defining "last expression"
usefully is difficult, so the smarter semantics can't fix all the cases).