Lower CilkScope to MIR with runtime hints
This commit changes lowering from THIR to MIR for CilkScope
to emit hints to start and stop the runtime at the beginning and
end of the block respectively. It additionally adds intrinsics
representing starting and stopping the runtime, and during codegen
these intrinsics are assumed to be matched since they're only
emitted from CilkScope.

Source information for the intrinsic calls is a little imprecise since
it refers to the entire cilk_scope block, but that might be fine.