Fix bug in SyncsIter::next
SyncsIter::next previously returned None when the value it got from the
internal iterator was none. This was incorrect since the internal iterator
returns values from next() which might not be sync locations,
while SyncsIter::next's job is to filter the values from the internal iterator.
The solution was to fix this with find_map rather than and_then.