Add test for variable initialized in all spawned tasks
This commit adds a test that a variable is considered initialized when
at least one task will be conditionally spawned and all tasks initialize
that variable. This test currently fails since MaybeUninitializedVariables
is not particularly good at reasoning about what dataflow should be restored
at a sync in terms of distinguishing between a set of tasks where at least 1
will run compared to where 0 may run and be synced.