Represent cilk_for in THIR and MIR
This commit adds a field is_parallel_loop_header to BasicBlockData
representing whether the basic block is a header to a Tapir parallel
loop. Its value is computed based on whether the loop corresponds to
a cilk_for, in which case the header of the loop has is_parallel_loop_header
set to true. This will be used later in lowering from MIR to LLVM IR to emit
the correct metadata. Whether the loop corresponds to a cilk_for is maintained
through thir::ExprKind::Loop::tapir_loop_spawn, which is true when the loop
corresponds to a cilk_for. ExprKind::Loop spawns the body when LoopSource
is CilkFor, so lowering to MIR is also responsible for syncing the body.