Lower parsed cilk_for to HIR by spawning body
This commit lowers parsed cilk_fors to HIR by slightly modifying
the standard for-loop lowering to loop {}. Instead of lowering to
loop { match next() { Some(..) => <body>, None => break } }, the body
is spawned (but only in the Some(..) case). This means forbidding
control flow in the body of a cilk_for should be sufficient. This
commit additionally removes the CilkSpawn introduced when lowering
HIR to THIR for parallel loops, since the body is already being
spawned.